% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSLSTM.R
\name{lagmatrix}
\alias{lagmatrix}
\title{Create lead/lags of a variable}
\usage{
lagmatrix(x, lag)
}
\arguments{
\item{x}{input variable.}

\item{lag}{vector of leads and lags. Positive numbers are lags, negative are leads. O is the original \code{x}.}
}
\value{
An array with the resulting leads and lags (columns).
}
\description{
Create an array with lead/lags of an input variable.
}
\note{
This code was copied from the \code{ts.utils} package to avoid the archive
operations of the \code{smooth} package in 16-02-2025. This function might be deprecated
in future releases to use the one from \code{ts.utils} again.
}
\examples{
  x <- rnorm(10)
  lagmatrix(x,c(0,1,-1))

}
\author{
Nikolaos Kourentzes, \email{nikolaos@kourentzes.com}
}
